+++ docs/components/DocumentationSidebarLink.tsx
@@ -69,7 +69,9 @@ type Props = {
   info: NavigationRoute;
 };
 
-export default class DocumentationSidebarLink extends React.Component<Props> {
+export default class DocumentationSidebarLink extends React.Component<
+  React.PropsWithChildren<Props>
+> {
   componentDidMount() {
     // Consistent link behavior across dev server and static export
     global.__NEXT_DATA__.nextExport = true;