+++ ios/versioned/sdk46/EXUpdatesInterface/EXUpdatesInterface/noop-file.m
@@ -1,0 +1,1 @@
+// noop file for cocoapods to generated modulemap file